uhosts quote fix to negate ! expansion

+++ src/st/hb.h
@@ -3,5 +3,5 @@
 #include <hb-ft.h>
 
 void hbunloadfonts();
-void hbtransform(XftGlyphFontSpec *, const Glyph *, int, int, int);
+void hbtransform(XftGlyphFontSpec *, const Glyph *, size_t, int, int);
